fix(mason): missing paths preventing some components from starting
- Mason now activates on `BufNewFile` / `BufReadPost` instead of by
running `:Mason`
- Automatic package installation now runs when the Mason UI opens
instead of being triggered on Mason's activation
- The custom `vim.env.PATH` handling is no longer maintained
